requirements.txt
asgiref==3.8.1
Django==4.2.1
sqlparse==0.5.1
django-suit==2.0a2